OS/10. Software Development OS/Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
mistralai==0.0.12
pillow==10.1.0
psycopg2-binary==2.9.9
redis==5.0.1
rq==1.15.1
pgvector==0.2.4
pyyaml==6.0.1
agno==0.1.0
python-multipart==0.0.6
requests==2.31.0
